2. Aqlan Research Agent - Research and Benchmarking/BackEnd/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.1.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
eval_type_backport==0.2.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
google_search_results==2.4.2
gpt4all==2.8.2
groq==0.12.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipython==8.12.3
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-mistralai==0.2.1
langchain-nomic==0.1.4
langchain-ollama==0.2.0
langchain-openai==0.2.8
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langgraph==0.2.48
langgraph-checkpoint==2.0.4
langgraph-sdk==0.1.36
langsmith==0.1.143
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.15
mistralai==1.2.2
mistune==3.0.2
ml_dtypes==0.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.2
nomic==3.3.3
numpy==1.26.4
ollama==0.3.3
openai==1.54.4
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pipreqs==0.5.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-router==0.0.72
sentence-transformers==3.3.0
serpapi==0.1.5
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
sympy==1.13.1
tavily-python==0.5.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
torch==2.5.1
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
xxhash==3.5.0
yarg==0.1.9
yarl==1.17.1
